Separate stack allocator into a dedicated namespaec
